+++ dep/g3dlite/include/G3D/Vector2int16.h
@@ -143,7 +143,7 @@ typedef Vector2int16 Point2int16;
 }
 
 template<> struct HashTrait<G3D::Vector2int16> {
-    static size_t hashCode(const G3D::Vector2int16& key) { return static_cast<size_t>(key.x + ((int)key.y << 16)); }
+    static size_t hashCode(const G3D::Vector2int16& key) { return static_cast<size_t>((int)key.x + ((int)key.y << 16)); }
 };
 
 #ifdef _MSC_VER
